+++ packages/frontend/src/pages/Periods/ReceiverSummary/components/ReceiverSummaryTable.tsx
@@ -1,14 +1,10 @@
-import { ForwarderTooltip } from '@/components/praise/ForwarderTooltip';
+import { PraiseDetailsDto } from 'api/dist/praise/types';
+import { useHistory, useParams } from 'react-router-dom';
 import Praise from '@/components/praise/Praise';
-import { UserAvatar } from '@/components/user/UserAvatar';
 import {
   PeriodAndReceiverPageParams,
   usePeriodReceiverPraiseQuery,
 } from '@/model/periods';
-import { localizeAndFormatIsoDate } from '@/utils/date';
-import { PraiseDetailsDto } from 'api/dist/praise/types';
-import { useHistory, useParams } from 'react-router-dom';
-
 interface PraiseRowProps {
   praise: PraiseDetailsDto;
 }
@@ -25,9 +21,6 @@ const PraiseRow = ({ praise }: PraiseRowProps): JSX.Element => {
       onClick={handleClick}
     >
       <Praise praise={praise} showReceiver={false} />
-      <div className="flex-grow text-right px-14 whitespace-nowrap">
-        {praise.scoreRealized}
-      </div>
     </div>
   );
 };